+++ programs/dark-bonds/src/state/ibo.rs
@@ -10,7 +10,7 @@ pub struct Ibo {
     pub fixed_exchange_rate: u64,
 
     // TODO option nto to enable swaps at all?
-    pub swap_cut: u32, // in % x 100
+    pub swap_cut: u64, // in % x 100
 
     // Can be purchased after this date
     pub live_date: i64,
